PermMissingElem problem solved

+++ java/src/lessons/three/complexity/PermMissingElem.java
@@ -7,12 +7,13 @@ public class PermMissingElem {
 
 	public int solution(int[] A) {
 		int N = A.length + 1;
-		int sumN = (1 + N) * N / 2;
+		double sumN = (1.0 + N) * N / 2.0 ;
 		
 		for (int i = 0; i < A.length; i++) {
 			sumN -= A[i];
 		}
 		
-		return sumN;
+		return (int) sumN;
 	}
+	
 }
